Muted composer question, fixed apache log paths
Composer warning was about running as root.
Muted the warning/prompt from showing, adding the --no-plugins flag for
a little extra safety.
AFAICT, the composer warning shows as since the top-level project's
(BookStack in this case) scripts will be ran as root in this case.
Since the user is already running the whole script as root they'd
already be putting trust in this install process so preventing running
scripts as the root user won't achieve a massive amount.